Rename NativeActivityIndicator to MauiActivityIndicator
