+++ src/Core/src/Handlers/ActivityIndicator/ActivityIndicatorHandler.iOS.cs
@@ -3,9 +3,9 @@ using UIKit;
 
 namespace Microsoft.Maui.Handlers
 {
-	public partial class ActivityIndicatorHandler : AbstractViewHandler<IActivityIndicator, NativeActivityIndicator>
+	public partial class ActivityIndicatorHandler : AbstractViewHandler<IActivityIndicator, MauiActivityIndicator>
 	{
-		protected override NativeActivityIndicator CreateNativeView() => new NativeActivityIndicator(CGRect.Empty, VirtualView)
+		protected override MauiActivityIndicator CreateNativeView() => new MauiActivityIndicator(CGRect.Empty, VirtualView)
 		{
 			ActivityIndicatorViewStyle = UIActivityIndicatorViewStyle.Gray
 		};
